react: Detect automatically if the webchat is running in iOS/Android webview (#3050)
## Description
Set the default target parameter in links depending if the webchat is
opened within a iOS/Android webview (`_self`) or not (`_blank`)

## Context
If the webchat is displayed within an iOS webview (inside an iOS App),
the links weren't working because the default target was `_blank` and on
iOS webviews the links only work if the target is `_self`.

## Approach taken / Explain the design
Create a new `isInWebView` function to detect if the environment is a
webview in iOS/Android.

## Testing

The pull request has no tests.