fix(react): Fix type of user argument in updateUser function (#3055)
## Description
Set the type of the `user` argument of the `updateUser` function as
`Partial<ClientUser>` because you may want to update just part of the
user attributes, not all of them.

## Context
If you wanted to update just some user attributes, you would get a
Typescript error.

## Testing

The pull request has no tests.