react & plugin-flow-builder: custom rating message #BLT-1746  (#3066)
## Description

**@botonic/react:**
Create CustomRatingMessage. This component receives the message text,
the payloads, the text of the send button and ratingType to know whether
to use stars or smileys.

**@botonic/plugin-flow-builder:**
Adds a new setting in the configuration that allows to activate the
custom rating. If it is activated and you have to render a rating in
webchat use the new CustomRatingMessage

## Context

The send button is disabled until the user has selected a number of
stars or a face in order to be able to submit the rating.
When the user clicks on the send button, the information in the
messageJSON of the localStorage is updated so that if the browser is
reloaded, the message comes out again as answered and the button is
disabled.

To edit styles for this custom message you can add a css, the component
has an empty css classes to facilitate changes.
The classes are: `custom-rating`, `rating-message-container` and
`rating-message-bubble`.
This is the css for the example you can see above:

```css
.custom-rating {
  width: 100% !important;
  .rating-message-container {
    .rating-message-bubble {
      background: $neutral-50 !important;
      border-radius: 10px 10px 10px 0px !important;
      color: $black !important;
      font-weight: 400 !important;
      padding: 12px !important;

      svg {
        color: $lilara-500 !important;
      }
    }
  }

  button {
      display: flex;
      justify-content: center;
      align-items: center;
      box-sizing: border-box;

      width: 100%;
      height: 37px;
      padding: 10px 16px;
      margin: 0px !important;
      background: $white !important;
      border-radius: 0px 10px 10px 10px !important;
      border: 1px solid $lilara-500 !important;
      white-space: pre-line;
      color: $lilara-500 !important;
      cursor: pointer;
  
     &:hover {
       background-color: $lilara-500 !important;
       color: $white !important;
    }
    &:disabled {
      cursor: auto;
      background-color: $neutral-100 !important;
      color: $black !important;
    }
  }
}

```

## To document / Usage example

Stars
<img width="349" height="186" alt="Captura de pantalla 2025-07-21 a las
9 11 27"
src="https://github.com/user-attachments/assets/6172fb83-b33e-4391-a8a9-ed825840e728"
/>
<img width="345" height="146" alt="Captura de pantalla 2025-07-21 a las
9 12 10"
src="https://github.com/user-attachments/assets/c521ad4e-f9b5-4ee9-9ced-294a39953931"
/>

Smileys

<img width="343" height="189" alt="Captura de pantalla 2025-07-22 a las
11 42 32"
src="https://github.com/user-attachments/assets/4e1d84f4-ca11-41c3-917d-170dbba94bd9"
/>
<img width="353" height="157" alt="Captura de pantalla 2025-07-22 a las
11 42 42"
src="https://github.com/user-attachments/assets/cafcfcf7-e971-434d-9d3a-ca9a66fd7c43"
/>


## Testing

The pull request...

- has unit tests
- has integration tests
- doesn't need tests because... **[provide a description]**