botonic-react & plugin-flow-builder: WhatsAppCtaUrl accept document, video and image in header #BLT-1783 (#3074)
## Description

Add the possibility to send an image, video or document in the header of
the message. A text can still be sent in the header. Only one of these
types can be added as a header, it is not possible to combine more than
one.

## Testing

Update and add tests for WhatsAppCTAURLButton component in
@botonic/react and @botonic/plugin-flow-builder