plugin-ai-agents & plugin-flow-builder: create InferenceResponse for ai-agents #BLT-1803 (#3083)
## Description

plugin-ai-agents return an object with interface InferenceResponse. With
this object we can know the tools executed or if ai agents not respond
(exit is true). In the future we include in this response the guardrails
info.

```typescript
export type InferenceResponse = {
  messages: AgenticOutputMessage[]
  toolsExecuted: string[]
  exit: boolean
  inputGuardrailTriggered: boolean
  outputGuardrailTriggered: boolean
} | undefined
```

plugin -flow-builder use this object to generate response, in the future
PR use this info to track ai-agents responses

## Test

Update mockAiAgentResponse to pass tests