FlowBuilder: Export function to obtain the rating value selected (#3094)
## Description
Extract the logic that obtains the submitted rating information into a
function so it can be exported and used in the bots.

## Context
If the bot needs to do some extra logic with the rating selected, it
needs to have a way to obtain the rating value selected by the user.

## Approach taken / Explain the design

## To document / Usage example
After exporting this new `getRatingButtonClicked` function, we could add
a Plugin in the bot to do some extra logic in the `pre` method when a
rating is submitted:
```ts
import { AGENT_RATING_PAYLOAD, getRatingSubmittedInfo } from '@botonic/plugin-flow-builder'
...
async pre(request: PluginPreRequest<BotPlugins>) {
    const { payload, cmsPlugin } = getRequestData(request)
    const agentRatingPayloadRegex = new RegExp(`^${AGENT_RATING_PAYLOAD}.*`)

    if (agentRatingPayloadRegex.test(payload)) {
      const { value } = cmsPlugin.getRatingSubmittedInfo(payload)
      // Do some logic with this rating value
    }
  }
```

## Testing

The pull request has no tests