fix(plugin-flow-builder): Fix Plugin-flow-builder getValueFromKeyPath to access input or session objects (#3116)
## Description
Update the `getValueFromKeyPath` function to check for variables inside
the `input` or `session` objects if the key starts with `"input."` or
`"session."` (with the dot at the end) rather than only with `"input"`
or `"session"`.

## Context
When checking a variable named `inputReceived` (that exists in
`session.user.extra_data`) in a Custom Conditional in Flow Builder , the
Plugin was checking the `input` object rather than the `extra_data` one.

## Testing

The pull request has unit tests