plugin-flow-builder: send new events in for bot-action, conditionals, redirect and webview triggered  #BLT-2026 (#3129)
## Context

We want to track nodes such as Conditional and GoToFlow. The problem is
that, as the plugin was currently working, these nodes were not being
rendered as content because they did not render anything. The solution
is to create content that does not render anything so that it can be
tracked in FowBuilderAction.

Review commit by commit for more context.

## Description

- Add abstract method trackFlow in ContentFieldsBase and apply it in
every flow content.
- Create conditional nodes as a flow content to receive them in contents
array in FlowBuilderAction so they can be tracked before render.
- Create a FlowFactory to extract lines from the src/index.ts file and
encapsulate the logic of calling the fromHubtypeCMS method for each
content type.
- Send new events for GoToFlow and Channel, Country, Custom and
QueueStatus conditionals


## Approach taken / Explain the design

Standardize all nodes with trackFlow and toBotonic funcion. 
When FlowBuilderAction getContents obtain nodes as conditionals and
gotoflow and can call trackFlow function for each content.
Contents as goToFlow or conditionals in toBotonic function render only a
fragment `<></>` this is equivalent to no render.

## Testing

- Fix all tests, now when check contents we have more contents because
in array contents have conditionals and gotToFlow.
- Update conditional queue mock 
- Add new tests to check tracking for each conditional, bot-action,
webview triggered and redirect

---------

Co-authored-by: vanbasten17 <marcrabat@gmail.com>