feat(plugin-flow-builder): Boolean condition of a bot variable conditional on a non boolean variable (#3132)
## Description
Allow a Bot variable conditional to execute a boolean condition over a
non boolean variable so it returns `true` when the variable exists and
have any value and returns `false` when it is `undefined` or `null`.

## Context
In Flow Builder you can't check if a variable has a value or not because
the conditions are checking only their values (string, number or
boolean), but using the boolean condition on non boolean variables
should check if it has a value (`true`) or not (`false`).

## Approach taken / Explain the design
@Iru89 In the future this could be reviewed and refactored to extract
this logic into a new condition type to check only if a variable exists
or not.

## To document / Usage example

## Testing

The pull request has integration tests.